moved everything from model to user_model and user_controller respectively, then activated listeners in main
