+++ javascripts/main.js
@@ -1,5 +1,5 @@
 'use strict';
-
+const user_controller = require('./user_controller');
 let firebase = require('firebase');
 let model = require('./model');
 const cred = require('./config/fbCreds');
@@ -7,3 +7,8 @@ let fbConfig = require('./config/fbConfig');
 
 const fbURL = "https://stealth-turmoil.firebaseio.com";
 const $ = require("jquery");
+
+
+$(document).ready(function() {
+  user_controller.activateListeners();
+});
